Moved test files to new subfolder. Updated checkout and description with routers for new functionality in components
